requirements.txt
#dj-database-url==0.3.0
#Django==1.8.1
#django-postgrespool==0.3.0

#psycopg2==2.6
#SQLAlchemy==1.0.4
#whitenoise==1.0.6
#requests==2.3.0

Django>1.11
requests>2.18
gunicorn==19.7.1
python-dateutil>2.3.0
icalendar==3.11.7

# iPython
